requirements.txt
Pillow==7.0.0
ImageHash==4.0